requirements.txt
# Base Framework
Django==4.2.0
django-debug-toolbar==4.2.0

# Image Processing
Pillow==10.1.0

# Authentication
django-allauth==0.58.2

# Forms
django-crispy-forms==2.1
crispy-bootstrap5==0.7

# File Management
django-cleanup==8.0.0

# UI Components
django-widget-tweaks==1.5.0

# Rich Text Editor
django-ckeditor==6.7.0

# Caching
django-redis==5.4.0

# Async Tasks
celery==5.3.6

# Charts
django-chartjs==2.3.0

# Pagination
django-pure-pagination==0.3.0

# Search
django-haystack==3.2.1
Whoosh==2.7.4

# Testing
pytest-django==4.7.0
factory-boy==3.3.0

# Development Tools
django-extensions==3.2.3
ipython==8.18.1

# Deployment
gunicorn==21.2.0
whitenoise==6.6.0
